# swagger.yaml
openapi: 3.0.0
info:
  title: EliteHub Hotel Management API
  description: API documentation for EliteHub Hotel Management System
  version: 2.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Reservation:
      type: object
      properties:
        guestId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        checkInDate:
          type: string
          format: date-time
        checkOutDate:
          type: string
          format: date-time
        numberOfGuests:
          type: integer
...
    Department:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive]
        headOfDepartment:
          type: string
          format: uuid
        budget:
          type: number
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
...
    CreateDepartmentRequest:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
          example: "Security"
        code:
          type: string
          example: "security"
        description:
          type: string
        status:
          type: string
          enum: [active, inactive]
        headOfDepartment:
          type: string
          format: uuid
        budget:
          type: number
        contactEmail:
          type: string
        contactPhone:
          type: string
security:
  - bearerAuth: []
paths:
...
  /departments:
    get:
      summary: Get all departments
      description: Retrieve a list of all hotel departments.
      tags: [Departments]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
        - in: query
          name: includeInactive
          schema:
            type: boolean
      responses:
        200:
          description: List of departments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Department"
    post:
      summary: Create new department (Admin only)
      description: Create a new hotel department. Restricted to Admin role.
      tags: [Departments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        201:
          description: Department created
        403:
          description: Admin access required
        409:
          description: Department already exists
...
  /reservations:
    get:
      summary: Get all reservations (Reception/Management)
      tags: [Reservations]
      parameters:
        - in: query
          name: checkInDate
          schema:
            type: string
            format: date
      responses:
        200:
          description: List of reservations
    post:
      summary: Create reservation (Reception)
      tags: [Reservations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reservation"
      responses:
        201:
          description: Reservation created
...
  /users:
    get:
      summary: Get all users (Admin only)
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of users retrieved successfully
...